fix: gRPC StopAnalysis 세션 ID long 손실 + 응답 DTO 정수 타입 일관성
- ExerciseAnalysisService.stopAnalysis 에서 .intValue() 로 인한
  세션 ID 64bit -> 32bit wrap-around 가능성 제거
- ExercisesResponseDto/SessionUpdateResponseDto 의 sessionId/exerciseId
  를 Long 으로 통일하여 DB(bigint) -> 응답 경로 일관성 확보
- gRPC 이전 후 호출자가 사라진 /internal/exercises/pose-data 컨트롤러와
  대응 DTO 제거 (PoseDataService 시그니처와 불일치로 컴파일 차단 상태였음)

+++ backend/src/main/java/com/shadowfit/controller/ExercisesController.java
@@ -64,8 +64,8 @@ public class ExercisesController {
 
         // 응답 DTO 생성
         ExercisesResponseDto response = ExercisesResponseDto.builder()
-                .sessionId(sessionId.intValue())
-                .exerciseId(dto.getExerciseId().intValue())
+                .sessionId(sessionId)
+                .exerciseId(dto.getExerciseId())
                 .startTime(LocalDateTime.now())
                 .status(Status.IN_PROGRESS)
                 .build();
@@ -92,7 +92,7 @@ public class ExercisesController {
 
         // 응답 생성
         SessionUpdateResponseDto response = SessionUpdateResponseDto.builder()
-                .sessionId(sessionId.intValue())
+                .sessionId(sessionId)
                 .status(Status.COMPLETED)
                 .endTime(LocalDateTime.now())
                 .build();
